separate lasso_test from selector_utils
