+++ js/src/lasso_test.js
@@ -31,8 +31,8 @@
  * SOFTWARE.
  */
 
-// checks if a point is in lasso
 function point_in_lasso(point, vertices) {
+    // Checks if a point is in lasso
     var xi, xj, yi, yj, intersect,
         x = point[0], y = point[1], is_inside = false;
 
@@ -47,56 +47,6 @@ function point_in_lasso(point, vertices) {
     return is_inside;
 }
 
-function point_in_rectangle(point, x, y) {
-    // checks whether `point` is within the rectangle of coordinates
-    // (x0, y0) (x1, y0) (x1, y1) (x0, y1)
-    // If one of x or y is undefined, treat them as [-inf, inf]
-    
-    if (x.length == 0 && y.length == 0) { return false; }
-
-    var is_inside = true;
-    x.sort(function(a, b){return a-b});
-    y.sort(function(a, b){return a-b});
-
-    if (x.length != 0) {
-        is_inside = is_inside && x[0] <= point[0] && point[0] <= x[1];
-    }
-    if (y.length != 0) {
-        is_inside = is_inside && y[0] <= point[1] && point[1] <= y[1];
-    } 
-    return is_inside;
-}
-
-// checks whether two rectangles intersect
-function rect_inter_rect(x0, y0, x1, y1) {
-    return seg_inter_seg(x0, x1) && seg_inter_seg(y0, y1);
-}
-
-function lasso_inter_rect(x, y, vertices) {
-    // checks whether the lasso intersects the rectangle of coordinates
-    // (x0, y0) (x1, y0) (x1, y1) (x0, y1)
-
-    for (var i = 0; i < vertices.length; i++) {
-        if (point_in_rectangle(vertices[i], x, y)) { return true; }
-    }
-    return false;
-}
-
-function seg_inter_seg(p, q) {
-    // Checks whether the 1d segments [p0, p1] and [q0, q1] intersect
-    // If one of the segments is empty, treat it as [-inf, inf]
-    if (p.length == 0 || q.length == 0) {
-        return (p.length != 0 || q.length != 0);
-    }
-    p.sort(function(a, b){return a-b});
-    q.sort(function(a, b){return a-b});
-    return ((p[0] < q[0] != p[0] < q[1]) || (p[1] < q[0] != p[1] < q[1]) ||
-            (q[0] < p[0] != q[0] < p[1]) || (q[1] < p[0] != q[1] < p[1]));
-}
-
 module.exports = {
     point_in_lasso: point_in_lasso,
-    point_in_rectangle: point_in_rectangle,
-    rect_inter_rect: rect_inter_rect,
-    lasso_inter_rect: lasso_inter_rect,
 }
